[Refactor] moved ProtocolShared.hpp contents to ProtocolUtils.hpp along with createMessage and parseCmd
