+++ src/server/Protocols/GameProtocol.cpp
@@ -22,7 +22,7 @@ bool GameProtocol::waitForClients() {
     while (this->_connectedClients.size() < targetClientCount && !timeout.isExpired()) {
         while ((msg = this->_incomingMQ->pop())) {
             // Parse command
-            auto parsedCmd = this->parseCommand(*msg);
+            auto parsedCmd = ProtocolUtils::parseCommand(*msg);
 
             // if command invalid continue
             if (!parsedCmd)
@@ -140,7 +140,7 @@ void GameProtocol::sendDelEntity(EntityID id) const {
     ss << id;
 
     for (auto conn : this->_connectedClients)
-        this->_outgoingMQ->push(this->createMessage("DEL_ENT", ss.str(), conn.addr, conn.port));
+        this->_outgoingMQ->push(ProtocolUtils::createMessage("DEL_ENT", ss.str(), conn.addr, conn.port));
 }
 
 template <class T>
@@ -165,39 +165,6 @@ void GameProtocol::sendDelComponent(EntityID id, Connection client) const {
 // UTILITIES
 //
 //
-
-std::optional<ParsedCmd> GameProtocol::parseCommand(Message<std::string> msg) {
-    std::optional<ParsedCmd> output;
-    auto splitMsg = Utilities::splitStr(msg.getMsg(), " ");
-
-    // Error handling
-    if (splitMsg.size() != 2)
-        return {};
-    // Check for CRLF
-    if (splitMsg[1][splitMsg.size() - 1] != '\n' || splitMsg[1][splitMsg.size() - 2] != '\r')
-        return {};
-    splitMsg[1].erase(splitMsg[1].length() - 2, 2);
-
-    // Get command type
-    if (splitMsg[0] == "HERE")
-        output->cmd = Command::Here; // Note make sure you do ELSE IF
-    else
-        return {};
-
-    // Split args up in args and subargs
-    auto splitArgs = Utilities::splitStr(splitMsg[1], ";");
-    for (auto arg : splitArgs)
-        output->args.push_back(Utilities::splitStr(arg, ","));
-
-    return output;
-}
-
-Message<std::string> GameProtocol::createMessage(std::string cmd, std::string args, asio::ip::address addr, asio::ip::port_type port) {
-    std::string body = cmd;
-    body += " " + args + "\r\n";
-
-    return Message<std::string>(body, addr, port);
-}
 
 int GameProtocol::getPlayer(asio::ip::address addr, asio::ip::port_type port) {
     int i;
